requirements.txt
flask==3.1.0
python-docx==1.1.2
docx2txt==0.8
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
